+++ package.json
@@ -55,7 +55,7 @@
 		"types": "tsc --noEmit"
 	},
 	"dependencies": {
-		"@prismicio/react": "^3",
+		"@prismicio/react": "^3.3.0",
 		"@prismicio/simulator": "^0.2.0",
 		"imgix-url-builder": "^0.0.5",
 		"lz-string": "^1.5.0"
