+++ src/PrismicImage.tsx
@@ -1,23 +1,30 @@
 "use client";
 
+import { imgixLoader } from "./imgixLoader";
+import { devMsg } from "./lib/devMsg";
+import { resolveDefaultExport } from "./lib/resolveDefaultExport";
+import type { ImageFieldImage } from "@prismicio/client";
+import { isFilled } from "@prismicio/client";
+import type { ImgixURLParams } from "imgix-url-builder";
+import { buildURL } from "imgix-url-builder";
+import type { ImageProps } from "next/image";
+import Image from "next/image";
 import type {
 	ForwardRefExoticComponent,
 	PropsWithoutRef,
-	RefAttributes} from "react";
-import {
-	forwardRef
+	RefAttributes,
 } from "react";
-import type { ImageProps } from "next/image";
-import Image from "next/image";
-import type { ImgixURLParams } from "imgix-url-builder";
-import { buildURL } from "imgix-url-builder";
-import type { ImageFieldImage} from "@prismicio/client";
-import { isFilled } from "@prismicio/client";
+import { forwardRef } from "react";
 
-import { devMsg } from "./lib/devMsg";
-import { resolveDefaultExport } from "./lib/resolveDefaultExport";
+/**
+ * @deprecated Use `PrismicImage` instead.
+ */
+export { PrismicImage as PrismicNextImage };
 
-import { imgixLoader } from "./imgixLoader";
+/**
+ * @deprecated Use `PrismicImageProps` instead.
+ */
+export type { PrismicImageProps as PrismicNextImageProps };
 
 const castInt = (input: string | number | undefined): number | undefined => {
 	if (typeof input === "number" || typeof input === "undefined") {
@@ -33,10 +40,7 @@ const castInt = (input: string | number | undefined): number | undefined => {
 	}
 };
 
-export type PrismicImageProps = Omit<
-	ImageProps,
-	"src" | "alt" | "loader"
-> & {
+export type PrismicImageProps = Omit<ImageProps, "src" | "alt" | "loader"> & {
 	/** The Prismic Image field or thumbnail to render. */
 	field: ImageFieldImage | null | undefined;
 
@@ -92,94 +96,92 @@ export type PrismicImageProps = Omit<
 // The type annotation is necessary to avoid a type reference issue.
 export const PrismicImage: ForwardRefExoticComponent<
 	PropsWithoutRef<PrismicImageProps> & RefAttributes<HTMLImageElement>
-> = forwardRef<HTMLImageElement, PrismicImageProps>(
-	function PrismicImage(
-		{
-			field,
-			imgixParams = {},
-			alt,
-			fallbackAlt,
-			fill,
-			width,
-			height,
-			fallback = null,
-			loader = imgixLoader,
-			...restProps
-		},
-		ref,
-	) {
-		if (process.env.NODE_ENV === "development") {
-			if (typeof alt === "string" && alt !== "") {
-				console.warn(
-					`[PrismicImage] The "alt" prop can only be used to declare an image as decorative by passing an empty string (alt="") but was provided a non-empty string. You can resolve this warning by removing the "alt" prop or changing it to alt="". For more details, see ${devMsg(
-						"alt-must-be-an-empty-string",
-					)}`,
-				);
-			}
-
-			if (typeof fallbackAlt === "string" && fallbackAlt !== "") {
-				console.warn(
-					`[PrismicImage] The "fallbackAlt" prop can only be used to declare an image as decorative by passing an empty string (fallbackAlt="") but was provided a non-empty string. You can resolve this warning by removing the "fallbackAlt" prop or changing it to fallbackAlt="". For more details, see ${devMsg(
-						"alt-must-be-an-empty-string",
-					)}`,
-				);
-			}
-		}
-
-		if (!isFilled.imageThumbnail(field)) {
-			return <>{fallback}</>;
-		}
-
-		const resolvedImgixParams = imgixParams;
-		for (const x in imgixParams) {
-			if (resolvedImgixParams[x as keyof typeof resolvedImgixParams] === null) {
-				resolvedImgixParams[x as keyof typeof resolvedImgixParams] = undefined;
-			}
-		}
-
-		const src = buildURL(field.url, imgixParams as ImgixURLParams);
-
-		const ar = field.dimensions.width / field.dimensions.height;
-
-		const castedWidth = castInt(width);
-		const castedHeight = castInt(height);
-
-		let resolvedWidth = castedWidth ?? field.dimensions.width;
-		let resolvedHeight = castedHeight ?? field.dimensions.height;
-
-		if (castedWidth != null && castedHeight == null) {
-			resolvedHeight = castedWidth / ar;
-		} else if (castedWidth == null && castedHeight != null) {
-			resolvedWidth = castedHeight * ar;
-		}
-
-		// A non-null assertion is required since we can't statically
-		// know if an alt attribute is available.
-		const resolvedAlt = (alt ?? (field.alt || fallbackAlt))!;
-
-		if (
-			process.env.NODE_ENV === "development" &&
-			typeof resolvedAlt !== "string"
-		) {
-			console.error(
-				`[PrismicImage] The following image is missing an "alt" property. Please add Alternative Text to the image in Prismic. To mark the image as decorative instead, add one of \`alt=""\` or \`fallbackAlt=""\`.`,
-				src,
+> = forwardRef<HTMLImageElement, PrismicImageProps>(function PrismicImage(
+	{
+		field,
+		imgixParams = {},
+		alt,
+		fallbackAlt,
+		fill,
+		width,
+		height,
+		fallback = null,
+		loader = imgixLoader,
+		...restProps
+	},
+	ref,
+) {
+	if (process.env.NODE_ENV === "development") {
+		if (typeof alt === "string" && alt !== "") {
+			console.warn(
+				`[PrismicImage] The "alt" prop can only be used to declare an image as decorative by passing an empty string (alt="") but was provided a non-empty string. You can resolve this warning by removing the "alt" prop or changing it to alt="". For more details, see ${devMsg(
+					"alt-must-be-an-empty-string",
+				)}`,
 			);
 		}
 
-		const ResolvedImage = resolveDefaultExport(Image);
+		if (typeof fallbackAlt === "string" && fallbackAlt !== "") {
+			console.warn(
+				`[PrismicImage] The "fallbackAlt" prop can only be used to declare an image as decorative by passing an empty string (fallbackAlt="") but was provided a non-empty string. You can resolve this warning by removing the "fallbackAlt" prop or changing it to fallbackAlt="". For more details, see ${devMsg(
+					"alt-must-be-an-empty-string",
+				)}`,
+			);
+		}
+	}
 
-		return (
-			<ResolvedImage
-				ref={ref}
-				src={src}
-				width={fill ? undefined : resolvedWidth}
-				height={fill ? undefined : resolvedHeight}
-				alt={resolvedAlt}
-				fill={fill}
-				loader={loader === null ? undefined : loader}
-				{...restProps}
-			/>
+	if (!isFilled.imageThumbnail(field)) {
+		return <>{fallback}</>;
+	}
+
+	const resolvedImgixParams = imgixParams;
+	for (const x in imgixParams) {
+		if (resolvedImgixParams[x as keyof typeof resolvedImgixParams] === null) {
+			resolvedImgixParams[x as keyof typeof resolvedImgixParams] = undefined;
+		}
+	}
+
+	const src = buildURL(field.url, imgixParams as ImgixURLParams);
+
+	const ar = field.dimensions.width / field.dimensions.height;
+
+	const castedWidth = castInt(width);
+	const castedHeight = castInt(height);
+
+	let resolvedWidth = castedWidth ?? field.dimensions.width;
+	let resolvedHeight = castedHeight ?? field.dimensions.height;
+
+	if (castedWidth != null && castedHeight == null) {
+		resolvedHeight = castedWidth / ar;
+	} else if (castedWidth == null && castedHeight != null) {
+		resolvedWidth = castedHeight * ar;
+	}
+
+	// A non-null assertion is required since we can't statically
+	// know if an alt attribute is available.
+	const resolvedAlt = (alt ?? (field.alt || fallbackAlt))!;
+
+	if (
+		process.env.NODE_ENV === "development" &&
+		typeof resolvedAlt !== "string"
+	) {
+		console.error(
+			`[PrismicImage] The following image is missing an "alt" property. Please add Alternative Text to the image in Prismic. To mark the image as decorative instead, add one of \`alt=""\` or \`fallbackAlt=""\`.`,
+			src,
 		);
-	},
-);
+	}
+
+	const ResolvedImage = resolveDefaultExport(Image);
+
+	return (
+		<ResolvedImage
+			ref={ref}
+			src={src}
+			width={fill ? undefined : resolvedWidth}
+			height={fill ? undefined : resolvedHeight}
+			alt={resolvedAlt}
+			fill={fill}
+			loader={loader === null ? undefined : loader}
+			{...restProps}
+		/>
+	);
+});
