--- src/deprecated.ts
@@ -1,19 +1,0 @@
-/**
- * @deprecated Use `PrismicImage` instead.
- */
-export { PrismicImage as PrismicNextImage } from "./PrismicImage";
-
-/**
- * @deprecated Use `PrismicImageProps` instead.
- */
-export type { PrismicImageProps as PrismicNextImageProps } from "./PrismicImage";
-
-/**
- * @deprecated Use `PrismicLink` instead.
- */
-export { PrismicLink as PrismicNextLink } from "./PrismicLink";
-
-/**
- * @deprecated Use `PrismicLinkProps` instead.
- */
-export type { PrismicLinkProps as PrismicNextLinkProps } from "./PrismicLink";
